+++ src/compiler/scala/reflect/macros/runtime/Infrastructure.scala
@@ -1,0 +1,16 @@
+package scala.reflect.macros
+package runtime
+
+trait Infrastructure {
+  self: Context =>
+
+  def settings: List[String] = {
+    val us = universe.settings
+    import us._
+    userSetSettings collectFirst { case x: MultiStringSetting if x.name == XmacroSettings.name => x.value } getOrElse Nil
+  }
+
+  def compilerSettings: List[String] = universe.settings.recreateArgs
+
+  def classPath: List[java.net.URL] = global.classPath.asURLs
+}
